+++ pages/_document.js
@@ -12,11 +12,7 @@ export default class MyDocument extends Document {
       <Html lang="pl">
         <Head>
           {/* Essential */}
-          <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-          <meta
-            name="viewport"
-            content="width=device-width,initial-scale=1,minimum-scale=1,maximum-scale=1,user-scalable=no"
-          />
+          <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
           <link href="/manifest.json" rel="manifest" />
 
           {/* Android */}