+++ pages/blog/index.js
@@ -1,3 +1,18 @@
-export default function Blog () {
-    return 'Coming soon ...'
+export default function Blog() {
+  return (
+    // <iframe
+    //   src="https://www.facebook.com/plugins/post.php?href=https%3A%2F%2Fwww.facebook.com%2F20531316728%2Fposts%2F10154009990506729%2F&show_text=true&width=552&height=274&appId"
+    //   width="552"
+    //   height="274"
+    //   style="border:none;overflow:hidden"
+    //   scrolling="no"
+    //   frameborder="0"
+    //   allowTransparency="true"
+    //   allow="encrypted-media"
+    // />
+    <h1 style={{
+      display: 'block',
+      margin: 'auto'
+    }}>Już wkrótce ... </h1>
+  );
 }